+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+### next [????.??.??]
+* Implement `qAddForeignFile` for `Quasi` instances
+
 ### 0.13.3 [2016.11.09]
 * Backport `Generic NameFlavour` instance